// libs/designer-v2/src/lib/ui/panel/recommendation/azureResourceSelection.d.ts
import type { DiscoveryOperation, DiscoveryResultTypes } from '@microsoft/logic-apps-shared';
interface AzureResourceSelectionProps {
    operation: DiscoveryOperation<DiscoveryResultTypes>;
}
export declare const AzureResourceSelection: (props: AzureResourceSelectionProps) => import("react/jsx-runtime").JSX.Element;
export {};