// libs/designer-v2/src/lib/ui/panel/recommendation/customSwaggerSelection.d.ts
import type { DiscoveryOperation, DiscoveryResultTypes } from '@microsoft/logic-apps-shared';
interface CustomSwaggerSelectionProps {
    operation: DiscoveryOperation<DiscoveryResultTypes>;
}
export declare const CustomSwaggerSelection: (props: CustomSwaggerSelectionProps) => import("react/jsx-runtime").JSX.Element;
export {};